Fix Events, update errors and update SimpleEvaluator

+++ src/ga/mutations/random_substitution.rs
@@ -51,7 +51,7 @@ impl<'a, R: Rng> Mutation for RandomSubstitutionMutation<'a, R> {
             // 50% chance to either extend the list of resources assigned to the event,
             // or to substitute for another resource instead.
             let chance = if resources.len() > 0 {
-                1.0 / (resources.len() + 1)
+                1.0 / (resources.len() as f64 + 1.0)
             } else {
                 0.0
             };
